requirements.txt
Flask==3.1.0
pytest
openai>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
